requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
Authlib==1.4.1
bcrypt==4.2.1
black==24.10.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.43.0
colorama==0.4.6
cryptography==3.3.2
cssbeautifier==1.15.4
Deprecated==1.2.18
djlint==1.36.4
dnspython==2.7.0
ecdsa==0.19.0
EditorConfig==0.17.0
email_validator==2.2.0
Flask==3.1.0
Flask-Authlib==1.6.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-simple-captcha==5.5.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsbeautifier==1.15.4
json5==0.12.0
jwt==1.3.1
limits==4.0.1
mailjet-rest==1.3.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
oauthlib==3.2.2
ordered-set==4.1.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.4
pytest-sugar==1.0.0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.37
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
pyodbc==5.2.0
pymssql==2.3.6